requirements.txt
aiohttp==3.7.4.post0
APScheduler==3.7.0
async-timeout==3.0.1
attrs==21.2.0
autopep8==1.5.7
cffi==1.14.6
chardet==4.0.0
discord-py-slash-command==2.3.2
discord.py==1.7.3
idna==3.2
multidict==5.1.0
pycodestyle==2.7.0
pycparser==2.20
PyNaCl==1.4.0
python-dotenv==0.19.0
pyttsx3==2.90
pytz==2021.1
six==1.16.0
SpeechRecognition==3.8.1       
toml==0.10.2
typing-extensions==3.10.0.0    
tzlocal==2.1
yarl==1.6.3